Init commit

+++ src/Vector2.java
@@ -1,8 +1,8 @@
 public class Vector2 {
-    public float x = 0;
-    public float y = 0;
+    public double x = 0;
+    public double y = 0;
 
-    public Vector2(float x, float y){
+    public Vector2(double x, double y){
         this.x = x;
         this.y = y;
     }
